refactor: book findOne, delete

+++ src/use-cases/book/deleteOne.js
@@ -1,6 +1,19 @@
 const serviceDeleteOne = (repository) => {
   return async (book_id) => {
-    await repository.deleteOne(book_id);
+    const user = await repository.findOne({ name: 'book_id', value: book_id });
+
+    if (!user) {
+      return {
+        status: false,
+        message: 'Book Not Found',
+      };
+    } else {
+      await repository.deleteOne(book_id);
+      return {
+        status: true,
+        message: 'Book has been deleted',
+      };
+    }
   };
 };
 module.exports = serviceDeleteOne;
